edit color and transform

+++ src/core/camera.rs
@@ -6,18 +6,18 @@ the shaders can then calc the right vertices
 zoom is how big units appear on the screen
 */
 pub struct Camera {
-    pub width: i32,
-    pub height: i32,
+    pub width: f32,
+    pub height: f32,
 }
 
 impl Camera {
     // create a new camera 
-    pub fn new(width: i32, height: i32) -> Camera {
+    pub fn new(width: f32, height: f32) -> Camera {
         Camera{width, height}
     }
 
     // refresh the cameras width and height, like after resizing sdl window
-    pub fn set_dim(&mut self, width: i32, height: i32) {
+    pub fn set_dim(&mut self, width: f32, height: f32) {
         self.width = width;
         self.height = height;
     }
